+++ example/a.ts
@@ -1,24 +1,26 @@
 import { Elysia, t } from 'elysia'
 import { treaty } from '../src'
 
-const app = new Elysia().get(
-    '/profile',
-    ({ headers }) => ({ user: 'hello', token: headers.authorization }),
-    {
-        // <- enforce that Authorization must be present
-        headers: t.Object({
-            authorization: t.String()
-        })
-    }
-)
+const app = new Elysia().get('/query', ({ query }) => query, {
+    query: t.Object({
+        orderBy: t.Array(
+            t.Object({
+                column: t.String()
+            })
+        )
+    })
+})
 
-const api = treaty(app, {
-    headers: {
-        authorization: ''
+const api = treaty(app)
+
+const { data } = await api.query.get({
+    query: {
+        orderBy: [
+            {
+                column: 'finalizedAt'
+            }
+        ]
     }
 })
-type api = typeof api
 
-const { status } = await api.profile.get({
-    headers: {}
-})
+console.log(data)